+++ test/e2e/test_deb.py
@@ -42,6 +42,14 @@ def deb_archive_with_cartridge(cartridge_cmd, tmpdir, project_with_cartridge):
                 "neofetch < 25")
 
     net_msg_max = 1024
+
+    systemd_unit_params = os.path.join(tmpdir, "systemd-unit-params.yml")
+    with open(systemd_unit_params, "w") as f:
+        f.write(f"""
+                instance-args:
+                    net_msg_max: {net_msg_max}
+                """)
+
     patch_init_to_net_msg_max(project, net_msg_max)
 
     cmd = [
@@ -50,7 +58,7 @@ def deb_archive_with_cartridge(cartridge_cmd, tmpdir, project_with_cartridge):
         "--deps-file", deps_filepath,
         "--preinst", pre_install_filepath,
         "--postinst", post_install_filepath,
-        "--net-msg-max", str(net_msg_max),
+        "--unit-params-file", systemd_unit_params,
         project.path,
         "--use-docker",
     ]