+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ``cartridge pack`` now uses the VERSION file from the
   ``TARANTOOL_SDK_PATH`` environment variable on building in Docker
 
+- Ability to specify ``TARANTOOL_*`` parameter in the systemd unit template
+  (command ``cartridge pack``) in the ``systemd-unit-params.yml`` file.
+
 ### Fixed
 
 - Improved tests in ``cartridge create`` template: